rename package to `diagnostic`
