+++ lsp-server/src/main/scala/org/alephium/ralph/lsp/server/converter/DiagnosticsConverter.scala
@@ -1,6 +1,6 @@
 package org.alephium.ralph.lsp.server.converter
 
-import org.alephium.ralph.lsp.pc.diagnostics.{CodeDiagnostic, CodeDiagnosticSeverity, FileDiagnostic}
+import org.alephium.ralph.lsp.pc.diagnostic.{CodeDiagnostic, CodeDiagnosticSeverity, FileDiagnostic}
 import org.eclipse.lsp4j._
 import scala.jdk.CollectionConverters.SeqHasAsJava
