+++ presentation-compiler/src/main/scala/org/alephium/ralph/lsp/pc/diagnostic/CodeDiagnosticSeverity.scala
@@ -1,4 +1,4 @@
-package org.alephium.ralph.lsp.pc.diagnostics
+package org.alephium.ralph.lsp.pc.diagnostic
 
 sealed trait CodeDiagnosticSeverity
