+++ presentation-compiler/src/main/scala/org/alephium/ralph/lsp/pc/diagnostic/FileDiagnostic.scala
@@ -1,4 +1,4 @@
-package org.alephium.ralph.lsp.pc.diagnostics
+package org.alephium.ralph.lsp.pc.diagnostic
 
 import java.net.URI
